requirements.txt
Django==2.2.4
django-cors-headers==3.0.2
django-filter==2.2.0
django-jalali==3.1.0
django-right-to-left==0.1.1
djangorestframework==3.10.2
jdatetime==3.5.0
Pillow==6.1.0
psycopg2==2.8.3
pytz==2019.2
sqlparse==0.3.0
